+++ partiql-lang/src/test/kotlin/org/partiql/lang/syntax/PartiQLParserCastTests.kt
@@ -12,6 +12,8 @@ import org.partiql.lang.util.ArgumentsProviderBase
 
 class PartiQLParserCastTests : PartiQLParserTestBase() {
 
+    override val targets: Array<ParserTarget> = arrayOf(ParserTarget.DEFAULT, ParserTarget.EXPERIMENTAL)
+
     @ParameterizedTest
     @ArgumentsSource(ConfiguredCastArguments::class)
     fun configuredCast(configuredCastCase: ConfiguredCastParseTest) = configuredCastCase.assertCase()